Adding Image #1
The image depicts the effect of height by position. The effect is quantified by looking at the slope for height on rebounds per game for position. The largest slope would indicate that an increase in height would be most beneficial for that position, while the smallest would indicate the opposite.